(main activity) created main activity and nav graph, updated presenter

+++ app/src/main/java/com/cs4520/assignment2/mvp/MVPModel.kt
@@ -1,19 +1,19 @@
 package com.cs4520.assignment2.mvp
 
 class MVPModel: Contract.Model {
-    override fun add(x:Float, y:Float): Float {
+    override fun add(x:Double, y:Double): Double {
         return x + y
     }
 
-    override fun subtract(x:Float, y:Float): Float {
+    override fun subtract(x:Double, y:Double): Double {
         return x - y
     }
 
-    override fun multiply(x:Float, y:Float): Float {
+    override fun multiply(x:Double, y:Double): Double {
         return x * y
     }
 
-    override fun divide(x:Float, y:Float): Float {
+    override fun divide(x:Double, y:Double): Double {
         return x / y
     }
 }
